splitting the code of GLBackend into separate .cpp, adding the PIpeline,adn the state to gpu
