+++ libraries/gpu/src/gpu/Batch.h
@@ -21,6 +21,8 @@
 #include "Stream.h"
 #include "Texture.h"
 
+#include "Pipeline.h"
+
 #if defined(NSIGHT_FOUND)
     #include "nvToolsExt.h"
     class ProfileRange {
@@ -96,7 +98,9 @@ public:
     void setViewTransform(const Transform& view);
     void setProjectionTransform(const Mat4& proj);
 
-    // Shader Stage
+    // Pipeline Stage
+    void setPipeline(const PipelinePointer& pipeline);
+
     void setUniformBuffer(uint32 slot, const BufferPointer& buffer, Offset offset, Offset size);
     void setUniformBuffer(uint32 slot, const BufferView& view); // not a command, just a shortcut from a BufferView
 
@@ -164,6 +168,7 @@ public:
         COMMAND_setViewTransform,
         COMMAND_setProjectionTransform,
 
+        COMMAND_setPipeline,
         COMMAND_setUniformBuffer,
         COMMAND_setUniformTexture,
 
@@ -281,6 +286,7 @@ public:
     typedef Cache<TexturePointer>::Vector TextureCaches;
     typedef Cache<Stream::FormatPointer>::Vector StreamFormatCaches;
     typedef Cache<Transform>::Vector TransformCaches;
+    typedef Cache<PipelinePointer>::Vector PipelineCaches;
 
     typedef unsigned char Byte;
     typedef std::vector<Byte> Bytes;
@@ -320,6 +326,7 @@ public:
     TextureCaches _textures;
     StreamFormatCaches _streamFormats;
     TransformCaches _transforms;
+    PipelineCaches _pipelines;
 
 protected:
 };
