+++ libraries/gpu/src/gpu/GLBackend.cpp
@@ -25,6 +25,7 @@ GLBackend::CommandCall GLBackend::_commandCalls[Batch::NUM_COMMANDS] =
     (&::gpu::GLBackend::do_setViewTransform),
     (&::gpu::GLBackend::do_setProjectionTransform),
 
+    (&::gpu::GLBackend::do_setPipeline),
     (&::gpu::GLBackend::do_setUniformBuffer),
     (&::gpu::GLBackend::do_setUniformTexture),
 
@@ -71,7 +72,8 @@ GLBackend::CommandCall GLBackend::_commandCalls[Batch::NUM_COMMANDS] =
 
 GLBackend::GLBackend() :
     _input(),
-    _transform()
+    _transform(),
+    _pipeline()
 {
     initTransform();
 }
@@ -134,6 +136,7 @@ void GLBackend::checkGLError() {
 void GLBackend::do_draw(Batch& batch, uint32 paramOffset) {
     updateInput();
     updateTransform();
+    updatePipeline();
 
     Primitive primitiveType = (Primitive)batch._params[paramOffset + 2]._uint;
     GLenum mode = _primitiveToGLmode[primitiveType];
@@ -147,6 +150,7 @@ void GLBackend::do_draw(Batch& batch, uint32 paramOffset) {
 void GLBackend::do_drawIndexed(Batch& batch, uint32 paramOffset) {
     updateInput();
     updateTransform();
+    updatePipeline();
 
     Primitive primitiveType = (Primitive)batch._params[paramOffset + 2]._uint;
     GLenum mode = _primitiveToGLmode[primitiveType];
@@ -166,389 +170,6 @@ void GLBackend::do_drawInstanced(Batch& batch, uint32 paramOffset) {
 void GLBackend::do_drawIndexedInstanced(Batch& batch, uint32 paramOffset) {
     CHECK_GL_ERROR();
 }
-
-void GLBackend::do_setInputFormat(Batch& batch, uint32 paramOffset) {
-    Stream::FormatPointer format = batch._streamFormats.get(batch._params[paramOffset]._uint);
-
-    if (format != _input._format) {
-        _input._format = format;
-        _input._invalidFormat = true;
-    }
-}
-
-void GLBackend::do_setInputBuffer(Batch& batch, uint32 paramOffset) {
-    Offset stride = batch._params[paramOffset + 0]._uint;
-    Offset offset = batch._params[paramOffset + 1]._uint;
-    BufferPointer buffer = batch._buffers.get(batch._params[paramOffset + 2]._uint);
-    uint32 channel = batch._params[paramOffset + 3]._uint;
-
-    if (channel < getNumInputBuffers()) {
-        _input._buffers[channel] = buffer;
-        _input._bufferOffsets[channel] = offset;
-        _input._bufferStrides[channel] = stride;
-        _input._buffersState.set(channel);
-    }
-}
-
-#define SUPPORT_LEGACY_OPENGL
-#if defined(SUPPORT_LEGACY_OPENGL)
-static const int NUM_CLASSIC_ATTRIBS = Stream::TANGENT;
-static const GLenum attributeSlotToClassicAttribName[NUM_CLASSIC_ATTRIBS] = {
-    GL_VERTEX_ARRAY,
-    GL_NORMAL_ARRAY,
-    GL_COLOR_ARRAY,
-    GL_TEXTURE_COORD_ARRAY
-};
-#endif
-
-void GLBackend::updateInput() {
-    if (_input._invalidFormat || _input._buffersState.any()) {
-
-        if (_input._invalidFormat) {
-            InputStageState::ActivationCache newActivation;
-
-            // Check expected activation
-            if (_input._format) {
-                const Stream::Format::AttributeMap& attributes = _input._format->getAttributes();
-                for (Stream::Format::AttributeMap::const_iterator it = attributes.begin(); it != attributes.end(); it++) {
-                    const Stream::Attribute& attrib = (*it).second;
-                    newActivation.set(attrib._slot);
-                }
-            }
-
-            // Manage Activation what was and what is expected now
-            for (unsigned int i = 0; i < newActivation.size(); i++) {
-                bool newState = newActivation[i];
-                if (newState != _input._attributeActivation[i]) {
-#if defined(SUPPORT_LEGACY_OPENGL)
-                    if (i < NUM_CLASSIC_ATTRIBS) {
-                        if (newState) {
-                            glEnableClientState(attributeSlotToClassicAttribName[i]);
-                        }
-                        else {
-                            glDisableClientState(attributeSlotToClassicAttribName[i]);
-                        }
-                    } else {
-#else 
-                    {
-#endif
-                        if (newState) {
-                            glEnableVertexAttribArray(i);
-                        } else {
-                            glDisableVertexAttribArray(i);
-                        }
-                    }
-                    CHECK_GL_ERROR();
-
-                    _input._attributeActivation.flip(i);
-                }
-            }
-        }
-
-        // now we need to bind the buffers and assign the attrib pointers
-        if (_input._format) {
-            const Buffers& buffers = _input._buffers;
-            const Offsets& offsets = _input._bufferOffsets;
-            const Offsets& strides = _input._bufferStrides;
-
-            const Stream::Format::AttributeMap& attributes = _input._format->getAttributes();
-
-            for (Stream::Format::ChannelMap::const_iterator channelIt = _input._format->getChannels().begin();
-                    channelIt != _input._format->getChannels().end();
-                    channelIt++) {
-                const Stream::Format::ChannelMap::value_type::second_type& channel = (*channelIt).second;
-                if ((*channelIt).first < buffers.size()) {
-                    int bufferNum = (*channelIt).first;
-
-                    if (_input._buffersState.test(bufferNum) || _input._invalidFormat) {
-                        GLuint vbo = gpu::GLBackend::getBufferID((*buffers[bufferNum]));
-                        glBindBuffer(GL_ARRAY_BUFFER, vbo);
-                        CHECK_GL_ERROR();
-                        _input._buffersState[bufferNum] = false;
-
-                        for (unsigned int i = 0; i < channel._slots.size(); i++) {
-                            const Stream::Attribute& attrib = attributes.at(channel._slots[i]);
-                            GLuint slot = attrib._slot;
-                            GLuint count = attrib._element.getDimensionCount();
-                            GLenum type = _elementTypeToGLType[attrib._element.getType()];
-                            GLuint stride = strides[bufferNum];
-                            GLuint pointer = attrib._offset + offsets[bufferNum];
-                            #if defined(SUPPORT_LEGACY_OPENGL)
-                            if (slot < NUM_CLASSIC_ATTRIBS) {
-                                switch (slot) {
-                                case Stream::POSITION:
-                                    glVertexPointer(count, type, stride, reinterpret_cast<GLvoid*>(pointer));
-                                    break;
-                                case Stream::NORMAL:
-                                    glNormalPointer(type, stride, reinterpret_cast<GLvoid*>(pointer));
-                                    break;
-                                case Stream::COLOR:
-                                    glColorPointer(count, type, stride, reinterpret_cast<GLvoid*>(pointer));
-                                    break;
-                                case Stream::TEXCOORD:
-                                    glTexCoordPointer(count, type, stride, reinterpret_cast<GLvoid*>(pointer)); 
-                                    break;
-                                };
-                            } else {
-                            #else 
-                            {
-                            #endif
-                                GLboolean isNormalized = attrib._element.isNormalized();
-                                glVertexAttribPointer(slot, count, type, isNormalized, stride,
-                                    reinterpret_cast<GLvoid*>(pointer));
-                            }
-                            CHECK_GL_ERROR();
-                        }
-                    }
-                }
-            }
-        }
-        // everything format related should be in sync now
-        _input._invalidFormat = false;
-    }
-
-/* TODO: Fancy version GL4.4
-    if (_needInputFormatUpdate) {
-
-        InputActivationCache newActivation;
-
-        // Assign the vertex format required
-        if (_inputFormat) {
-            const StreamFormat::AttributeMap& attributes = _inputFormat->getAttributes();
-            for (StreamFormat::AttributeMap::const_iterator it = attributes.begin(); it != attributes.end(); it++) {
-                const StreamFormat::Attribute& attrib = (*it).second;
-                newActivation.set(attrib._slot);
-                glVertexAttribFormat(
-                    attrib._slot,
-                    attrib._element.getDimensionCount(),
-                    _elementTypeToGLType[attrib._element.getType()],
-                    attrib._element.isNormalized(),
-                    attrib._stride);
-            }
-            CHECK_GL_ERROR();
-        }
-
-        // Manage Activation what was and what is expected now
-        for (int i = 0; i < newActivation.size(); i++) {
-            bool newState = newActivation[i];
-            if (newState != _inputAttributeActivation[i]) {
-                if (newState) {
-                    glEnableVertexAttribArray(i);
-                } else {
-                    glDisableVertexAttribArray(i);
-                }
-                _inputAttributeActivation.flip(i);
-            }
-        }
-        CHECK_GL_ERROR();
-
-        _needInputFormatUpdate = false;
-    }
-
-    if (_needInputStreamUpdate) {
-        if (_inputStream) {
-            const Stream::Buffers& buffers = _inputStream->getBuffers();
-            const Stream::Offsets& offsets = _inputStream->getOffsets();
-            const Stream::Strides& strides = _inputStream->getStrides();
-
-            for (int i = 0; i < buffers.size(); i++) {
-                GLuint vbo = gpu::GLBackend::getBufferID((*buffers[i]));
-                glBindVertexBuffer(i, vbo, offsets[i], strides[i]);
-            }
-
-            CHECK_GL_ERROR();
-        }
-        _needInputStreamUpdate = false;
-    }
-*/
-}
-
-
-void GLBackend::do_setIndexBuffer(Batch& batch, uint32 paramOffset) {
-    _input._indexBufferType = (Type) batch._params[paramOffset + 2]._uint;
-    BufferPointer indexBuffer = batch._buffers.get(batch._params[paramOffset + 1]._uint);
-    _input._indexBufferOffset = batch._params[paramOffset + 0]._uint;
-    _input._indexBuffer = indexBuffer;
-    if (indexBuffer) {
-        glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, getBufferID(*indexBuffer));
-    } else {
-        glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
-    }
-    CHECK_GL_ERROR();
-}
-
-// Transform Stage
-
-void GLBackend::do_setModelTransform(Batch& batch, uint32 paramOffset) {
-    _transform._model = batch._transforms.get(batch._params[paramOffset]._uint);
-    _transform._invalidModel = true;
-}
-
-void GLBackend::do_setViewTransform(Batch& batch, uint32 paramOffset) {
-    _transform._view = batch._transforms.get(batch._params[paramOffset]._uint);
-    _transform._invalidView = true;
-}
-
-void GLBackend::do_setProjectionTransform(Batch& batch, uint32 paramOffset) {
-    memcpy(&_transform._projection, batch.editData(batch._params[paramOffset]._uint), sizeof(Mat4));
-    _transform._invalidProj = true;
-}
-
-void GLBackend::initTransform() {
-#if defined(Q_OS_WIN)
-    glGenBuffers(1, &_transform._transformObjectBuffer);
-    glGenBuffers(1, &_transform._transformCameraBuffer);
-
-    glBindBuffer(GL_UNIFORM_BUFFER, _transform._transformObjectBuffer);
-    glBufferData(GL_UNIFORM_BUFFER, sizeof(_transform._transformObject), (const void*) &_transform._transformObject, GL_DYNAMIC_DRAW);
-
-    glBindBuffer(GL_UNIFORM_BUFFER, _transform._transformCameraBuffer);
-    glBufferData(GL_UNIFORM_BUFFER, sizeof(_transform._transformCamera), (const void*) &_transform._transformCamera, GL_DYNAMIC_DRAW);
-
-
-    glBindBuffer(GL_UNIFORM_BUFFER, 0);
-#else
-#endif
-}
-
-void GLBackend::killTransform() {
-#if defined(Q_OS_WIN)
-    glDeleteBuffers(1, &_transform._transformObjectBuffer);
-    glDeleteBuffers(1, &_transform._transformCameraBuffer);
-#else
-#endif
-}
-void GLBackend::updateTransform() {
-    // Check all the dirty flags and update the state accordingly
-    if (_transform._invalidProj) {
-        _transform._transformCamera._projection = _transform._projection;
-    }
-
-    if (_transform._invalidView) {
-        _transform._view.getInverseMatrix(_transform._transformCamera._view);
-        _transform._view.getMatrix(_transform._transformCamera._viewInverse);
-    }
-
-    if (_transform._invalidModel) {
-        _transform._model.getMatrix(_transform._transformObject._model);
-        _transform._model.getInverseMatrix(_transform._transformObject._modelInverse);
-    }
-
-    if (_transform._invalidView || _transform._invalidProj) {
-        Mat4 viewUntranslated = _transform._transformCamera._view;
-        viewUntranslated[3] = Vec4(0.0f, 0.0f, 0.0f, 1.0f);
-        _transform._transformCamera._projectionViewUntranslated = _transform._transformCamera._projection * viewUntranslated;
-    }
- 
-    if (_transform._invalidView || _transform._invalidProj) {
-#if defined(Q_OS_WIN)
-        glBindBufferBase(GL_UNIFORM_BUFFER, TRANSFORM_CAMERA_SLOT, 0);
-        glBindBuffer(GL_ARRAY_BUFFER, _transform._transformCameraBuffer);
-        glBufferData(GL_ARRAY_BUFFER, sizeof(_transform._transformCamera), (const void*) &_transform._transformCamera, GL_DYNAMIC_DRAW);
-        glBindBuffer(GL_ARRAY_BUFFER, 0);
-        CHECK_GL_ERROR();
-#endif
-   }
-
-    if (_transform._invalidModel) {
-#if defined(Q_OS_WIN)
-        glBindBufferBase(GL_UNIFORM_BUFFER, TRANSFORM_OBJECT_SLOT, 0);
-        glBindBuffer(GL_ARRAY_BUFFER, _transform._transformObjectBuffer);
-        glBufferData(GL_ARRAY_BUFFER, sizeof(_transform._transformObject), (const void*) &_transform._transformObject, GL_DYNAMIC_DRAW);
-        glBindBuffer(GL_ARRAY_BUFFER, 0);
-        CHECK_GL_ERROR();
-#endif
-    }
-
-#if defined(Q_OS_WIN)
-    glBindBufferBase(GL_UNIFORM_BUFFER, TRANSFORM_OBJECT_SLOT, _transform._transformObjectBuffer);
-    glBindBufferBase(GL_UNIFORM_BUFFER, TRANSFORM_CAMERA_SLOT, _transform._transformCameraBuffer);
-    CHECK_GL_ERROR();
-#endif
-
-
-#if defined(Q_OS_MAC) || defined(Q_OS_LINUX)
-    // Do it again for fixed pipeline until we can get rid of it
-    if (_transform._invalidProj) {
-        if (_transform._lastMode != GL_PROJECTION) {
-            glMatrixMode(GL_PROJECTION);
-            _transform._lastMode = GL_PROJECTION;
-        }
-        glLoadMatrixf(reinterpret_cast< const GLfloat* >(&_transform._projection));
-
-        CHECK_GL_ERROR();
-    }
-
-    if (_transform._invalidModel || _transform._invalidView) {
-        if (!_transform._model.isIdentity()) {
-            if (_transform._lastMode != GL_MODELVIEW) {
-                glMatrixMode(GL_MODELVIEW);
-                _transform._lastMode = GL_MODELVIEW;
-            }
-            Transform::Mat4 modelView;
-            if (!_transform._view.isIdentity()) {
-                Transform mvx;
-                Transform::inverseMult(mvx, _transform._view, _transform._model);
-                mvx.getMatrix(modelView);
-            } else {
-                _transform._model.getMatrix(modelView);
-            }
-            glLoadMatrixf(reinterpret_cast< const GLfloat* >(&modelView));
-        } else {
-            if (!_transform._view.isIdentity()) {
-                if (_transform._lastMode != GL_MODELVIEW) {
-                    glMatrixMode(GL_MODELVIEW);
-                    _transform._lastMode = GL_MODELVIEW;
-                }
-                Transform::Mat4 modelView;
-                _transform._view.getInverseMatrix(modelView);
-                glLoadMatrixf(reinterpret_cast< const GLfloat* >(&modelView));
-            } else {
-                // TODO: eventually do something about the matrix when neither view nor model is specified?
-                // glLoadIdentity();
-            }
-        }
-        CHECK_GL_ERROR();
-    }
-#endif
-
-    // Flags are clean
-    _transform._invalidView = _transform._invalidProj = _transform._invalidModel = false;
-}
-
-void GLBackend::do_setUniformBuffer(Batch& batch, uint32 paramOffset) {
-    GLuint slot = batch._params[paramOffset + 3]._uint;
-    BufferPointer uniformBuffer = batch._buffers.get(batch._params[paramOffset + 2]._uint);
-    GLintptr rangeStart = batch._params[paramOffset + 1]._uint;
-    GLsizeiptr rangeSize = batch._params[paramOffset + 0]._uint;
-#if defined(Q_OS_MAC)
-    GLfloat* data = (GLfloat*) (uniformBuffer->getData() + rangeStart);
-    glUniform4fv(slot, rangeSize / sizeof(GLfloat[4]), data);
- 
-    // NOT working so we ll stick to the uniform float array until we move to core profile
-    // GLuint bo = getBufferID(*uniformBuffer);
-    //glUniformBufferEXT(_shader._program, slot, bo);
-#elif defined(Q_OS_WIN)
-    GLuint bo = getBufferID(*uniformBuffer);
-    glBindBufferRange(GL_UNIFORM_BUFFER, slot, bo, rangeStart, rangeSize);
-#else
-    GLfloat* data = (GLfloat*) (uniformBuffer->getData() + rangeStart);
-    glUniform4fv(slot, rangeSize / sizeof(GLfloat[4]), data);
-#endif
-    CHECK_GL_ERROR();
-}
-
-void GLBackend::do_setUniformTexture(Batch& batch, uint32 paramOffset) {
-    GLuint slot = batch._params[paramOffset + 1]._uint;
-    TexturePointer uniformTexture = batch._textures.get(batch._params[paramOffset + 0]._uint);
-
-    GLuint to = getTextureID(uniformTexture);
-    glActiveTexture(GL_TEXTURE0 + slot);
-    glBindTexture(GL_TEXTURE_2D, to);
-
-    CHECK_GL_ERROR();
-}
-
 
 // TODO: As long as we have gl calls explicitely issued from interface
 // code, we need to be able to record and batch these calls. THe long 
@@ -740,8 +361,10 @@ void Batch::_glUseProgram(GLuint program) {
 }
 void GLBackend::do_glUseProgram(Batch& batch, uint32 paramOffset) {
     
-    _shader._program = batch._params[paramOffset]._uint;
-    glUseProgram(_shader._program);
+    _pipeline._program = batch._params[paramOffset]._uint;
+    // for this call we still want to execute the glUseProgram in the order of the glCOmmand to avoid any issue
+    _pipeline._invalidProgram = false;
+    glUseProgram(_pipeline._program);
 
     CHECK_GL_ERROR();
 }
@@ -996,57 +619,5 @@ void GLBackend::do_glColor4f(Batch& batch, uint32 paramOffset) {
         batch._params[paramOffset + 1]._float,
         batch._params[paramOffset + 0]._float);
     CHECK_GL_ERROR();
-}
-
-GLBackend::GLBuffer::GLBuffer() :
-    _stamp(0),
-    _buffer(0),
-    _size(0)
-{}
-
-GLBackend::GLBuffer::~GLBuffer() {
-    if (_buffer != 0) {
-        glDeleteBuffers(1, &_buffer);
-    }
-}
-
-GLBackend::GLBuffer* GLBackend::syncGPUObject(const Buffer& buffer) {
-    GLBuffer* object = Backend::getGPUObject<GLBackend::GLBuffer>(buffer);
-
-    if (object && (object->_stamp == buffer.getSysmem().getStamp())) {
-        return object;
-    }
-
-    // need to have a gpu object?
-    if (!object) {
-        object = new GLBuffer();
-        glGenBuffers(1, &object->_buffer);
-        CHECK_GL_ERROR();
-        Backend::setGPUObject(buffer, object);
-    }
-
-    // Now let's update the content of the bo with the sysmem version
-    // TODO: in the future, be smarter about when to actually upload the glBO version based on the data that did change
-    //if () {
-    glBindBuffer(GL_ARRAY_BUFFER, object->_buffer);
-    glBufferData(GL_ARRAY_BUFFER, buffer.getSysmem().getSize(), buffer.getSysmem().readData(), GL_DYNAMIC_DRAW);
-    glBindBuffer(GL_ARRAY_BUFFER, 0);
-    object->_stamp = buffer.getSysmem().getStamp();
-    object->_size = buffer.getSysmem().getSize();
-    //}
-    CHECK_GL_ERROR();
-
-    return object;
-}
-
-
-
-GLuint GLBackend::getBufferID(const Buffer& buffer) {
-    GLBuffer* bo = GLBackend::syncGPUObject(buffer);
-    if (bo) {
-        return bo->_buffer;
-    } else {
-        return 0;
-    }
 }
