+++ libraries/gpu/src/gpu/State.cpp
@@ -1,0 +1,20 @@
+//
+//  State.cpp
+//  libraries/gpu/src/gpu
+//
+//  Created by Sam Gateau on 3/8/2015.
+//  Copyright 2014 High Fidelity, Inc.
+//
+//  Distributed under the Apache License, Version 2.0.
+//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
+//
+
+#include "State.h"
+#include <QDebug>
+
+using namespace gpu;
+
+
+State::~State()
+{
+}
